Taks A: creating the Application
We created a development database.

> rails new depot
> cd depot

We created the products table and used the scaffold generator to write
an application to maintain it.

>rails generate scaffold Product title:string description:text
image_url:string price:decimal

We used a migration to create and modify the schema in our development
database.

> rake db:migrate
> rake db:seed

We updated an application-wide layout as well as a controller-specific
view in order to show a list of products.

app/assets/stylesheets/products.css.scss
app/views/layouts/application.html.erb
app/views/products/index.html.erb